pyproject.toml
[tool.poetry]
name = "cinema-tickets-api"
version = "0.1.0"
description = ""
authors = ["mina samaan <minalotfy05@gamil.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
pillow = "^10.4.0"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
